+++ instrumentation/github.com/Shopify/sarama/otelsarama/producer_test.go
@@ -26,9 +26,9 @@ import (
 
 	"go.opentelemetry.io/otel/codes"
 
-	"go.opentelemetry.io/otel/api/propagation"
 	"go.opentelemetry.io/otel/api/trace"
 	"go.opentelemetry.io/otel/label"
+	otelpropagators "go.opentelemetry.io/otel/propagators"
 	"go.opentelemetry.io/otel/semconv"
 
 	mocktracer "go.opentelemetry.io/contrib/internal/trace"
@@ -42,6 +42,7 @@ func NewTracerProviderAndTracer() (*mocktracer.TracerProvider, *mocktracer.Trace
 }
 
 func TestWrapSyncProducer(t *testing.T) {
+	propagators := otelpropagators.TraceContext{}
 	var err error
 
 	// Mock provider
@@ -52,12 +53,12 @@ func TestWrapSyncProducer(t *testing.T) {
 	mockSyncProducer := mocks.NewSyncProducer(t, cfg)
 
 	// Wrap sync producer
-	syncProducer := WrapSyncProducer(cfg, mockSyncProducer, WithTracerProvider(provider))
+	syncProducer := WrapSyncProducer(cfg, mockSyncProducer, WithTracerProvider(provider), WithPropagators(propagators))
 
 	// Create message with span context
 	ctx, _ := mt.Start(context.Background(), "")
 	messageWithSpanContext := sarama.ProducerMessage{Topic: topic, Key: sarama.StringEncoder("foo")}
-	propagation.InjectHTTP(ctx, propagators, NewProducerMessageCarrier(&messageWithSpanContext))
+	propagators.Inject(ctx, NewProducerMessageCarrier(&messageWithSpanContext))
 
 	// Expected
 	expectedList := []struct {
@@ -142,17 +143,18 @@ func TestWrapSyncProducer(t *testing.T) {
 		}
 
 		// Check tracing propagation
-		remoteSpanFromMessage := trace.RemoteSpanContextFromContext(propagation.ExtractHTTP(context.Background(), propagators, NewProducerMessageCarrier(msg)))
+		remoteSpanFromMessage := trace.RemoteSpanContextFromContext(propagators.Extract(context.Background(), NewProducerMessageCarrier(msg)))
 		assert.True(t, remoteSpanFromMessage.IsValid())
 	}
 }
 
 func TestWrapAsyncProducer(t *testing.T) {
+	propagators := otelpropagators.TraceContext{}
 	// Create message with span context
 	createMessages := func(mt *mocktracer.Tracer) []*sarama.ProducerMessage {
 		ctx, _ := mt.Start(context.Background(), "")
 		messageWithSpanContext := sarama.ProducerMessage{Topic: topic, Key: sarama.StringEncoder("foo")}
-		propagation.InjectHTTP(ctx, propagators, NewProducerMessageCarrier(&messageWithSpanContext))
+		propagators.Inject(ctx, NewProducerMessageCarrier(&messageWithSpanContext))
 		mt.EndedSpans()
 
 		return []*sarama.ProducerMessage{
@@ -167,7 +169,7 @@ func TestWrapAsyncProducer(t *testing.T) {
 
 		cfg := newSaramaConfig()
 		mockAsyncProducer := mocks.NewAsyncProducer(t, cfg)
-		ap := WrapAsyncProducer(cfg, mockAsyncProducer, WithTracerProvider(provider))
+		ap := WrapAsyncProducer(cfg, mockAsyncProducer, WithTracerProvider(provider), WithPropagators(propagators))
 
 		msgList := createMessages(mt)
 		// Send message
@@ -223,7 +225,7 @@ func TestWrapAsyncProducer(t *testing.T) {
 			}
 
 			// Check tracing propagation
-			remoteSpanFromMessage := trace.RemoteSpanContextFromContext(propagation.ExtractHTTP(context.Background(), propagators, NewProducerMessageCarrier(msg)))
+			remoteSpanFromMessage := trace.RemoteSpanContextFromContext(propagators.Extract(context.Background(), NewProducerMessageCarrier(msg)))
 			assert.True(t, remoteSpanFromMessage.IsValid())
 		}
 	})
@@ -237,7 +239,7 @@ func TestWrapAsyncProducer(t *testing.T) {
 		cfg.Producer.Return.Successes = true
 
 		mockAsyncProducer := mocks.NewAsyncProducer(t, cfg)
-		ap := WrapAsyncProducer(cfg, mockAsyncProducer, WithTracerProvider(provider))
+		ap := WrapAsyncProducer(cfg, mockAsyncProducer, WithTracerProvider(provider), WithPropagators(propagators))
 
 		msgList := createMessages(mt)
 		// Send message
@@ -300,13 +302,14 @@ func TestWrapAsyncProducer(t *testing.T) {
 			assert.Equal(t, i, msg.Metadata)
 
 			// Check tracing propagation
-			remoteSpanFromMessage := trace.RemoteSpanContextFromContext(propagation.ExtractHTTP(context.Background(), propagators, NewProducerMessageCarrier(msg)))
+			remoteSpanFromMessage := trace.RemoteSpanContextFromContext(propagators.Extract(context.Background(), NewProducerMessageCarrier(msg)))
 			assert.True(t, remoteSpanFromMessage.IsValid())
 		}
 	})
 }
 
 func TestWrapAsyncProducerError(t *testing.T) {
+	propagators := otelpropagators.TraceContext{}
 	// Mock provider
 	provider, mt := NewTracerProviderAndTracer()
 
@@ -315,7 +318,7 @@ func TestWrapAsyncProducerError(t *testing.T) {
 	cfg.Producer.Return.Successes = true
 
 	mockAsyncProducer := mocks.NewAsyncProducer(t, cfg)
-	ap := WrapAsyncProducer(cfg, mockAsyncProducer, WithTracerProvider(provider))
+	ap := WrapAsyncProducer(cfg, mockAsyncProducer, WithTracerProvider(provider), WithPropagators(propagators))
 
 	mockAsyncProducer.ExpectInputAndFail(errors.New("test"))
 	ap.Input() <- &sarama.ProducerMessage{Topic: topic, Key: sarama.StringEncoder("foo2")}
@@ -330,7 +333,7 @@ func TestWrapAsyncProducerError(t *testing.T) {
 
 	span := spanList[0]
 
-	assert.Equal(t, codes.Internal, span.Status)
+	assert.Equal(t, codes.Error, span.Status)
 	assert.Equal(t, "test", span.StatusMessage)
 }