+++ instrumentation/github.com/labstack/echo/otelecho/echo_test.go
@@ -28,12 +28,13 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/codes"
+	"go.opentelemetry.io/otel/propagators"
 
 	mocktrace "go.opentelemetry.io/contrib/internal/trace"
 	b3prop "go.opentelemetry.io/contrib/propagators/b3"
 	otelglobal "go.opentelemetry.io/otel/api/global"
-	otelpropagation "go.opentelemetry.io/otel/api/propagation"
 	oteltrace "go.opentelemetry.io/otel/api/trace"
 	"go.opentelemetry.io/otel/label"
 )
@@ -144,7 +145,7 @@ func TestError(t *testing.T) {
 	assert.Equal(t, label.IntValue(http.StatusInternalServerError), span.Attributes["http.status_code"])
 	assert.Equal(t, label.StringValue("oh no"), span.Attributes["echo.error"])
 	// server errors set the status
-	assert.Equal(t, codes.Internal, span.Status)
+	assert.Equal(t, codes.Error, span.Status)
 }
 
 func TestErrorOnlyHandledOnce(t *testing.T) {
@@ -182,12 +183,13 @@ func TestGetSpanNotInstrumented(t *testing.T) {
 
 func TestPropagationWithGlobalPropagators(t *testing.T) {
 	provider, tracer := mocktrace.NewTracerProviderAndTracer(tracerName)
+	otelglobal.SetTextMapPropagator(propagators.TraceContext{})
 
 	r := httptest.NewRequest("GET", "/user/123", nil)
 	w := httptest.NewRecorder()
 
 	ctx, pspan := tracer.Start(context.Background(), "test")
-	otelpropagation.InjectHTTP(ctx, otelglobal.Propagators(), r.Header)
+	otelglobal.TextMapPropagator().Inject(ctx, r.Header)
 
 	router := echo.New()
 	router.Use(Middleware("foobar", WithTracerProvider(provider)))
@@ -201,25 +203,22 @@ func TestPropagationWithGlobalPropagators(t *testing.T) {
 	})
 
 	router.ServeHTTP(w, r)
+	otelglobal.SetTextMapPropagator(otel.NewCompositeTextMapPropagator())
 }
 
 func TestPropagationWithCustomPropagators(t *testing.T) {
 	provider, tracer := mocktrace.NewTracerProviderAndTracer(tracerName)
 
 	b3 := b3prop.B3{}
-	props := otelpropagation.New(
-		otelpropagation.WithExtractors(b3),
-		otelpropagation.WithInjectors(b3),
-	)
 
 	r := httptest.NewRequest("GET", "/user/123", nil)
 	w := httptest.NewRecorder()
 
 	ctx, pspan := tracer.Start(context.Background(), "test")
-	otelpropagation.InjectHTTP(ctx, props, r.Header)
+	b3.Inject(ctx, r.Header)
 
 	router := echo.New()
-	router.Use(Middleware("foobar", WithTracerProvider(provider), WithPropagators(props)))
+	router.Use(Middleware("foobar", WithTracerProvider(provider), WithPropagators(b3)))
 	router.GET("/user/:id", func(c echo.Context) error {
 		span := oteltrace.SpanFromContext(c.Request().Context())
 		mspan, ok := span.(*mocktrace.Span)