+++ instrumentation/net/http/httptrace/otelhttptrace/httptrace.go
@@ -18,9 +18,8 @@ import (
 	"context"
 	"net/http"
 
-	"go.opentelemetry.io/otel/api/baggage"
+	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/api/global"
-	"go.opentelemetry.io/otel/api/propagation"
 	"go.opentelemetry.io/otel/api/trace"
 	"go.opentelemetry.io/otel/label"
 	"go.opentelemetry.io/otel/semconv"
@@ -31,11 +30,11 @@ import (
 type Option func(*config)
 
 type config struct {
-	propagators propagation.Propagators
+	propagators otel.TextMapPropagator
 }
 
 func newConfig(opts []Option) *config {
-	c := &config{propagators: global.Propagators()}
+	c := &config{propagators: global.TextMapPropagator()}
 	for _, o := range opts {
 		o(c)
 	}
@@ -43,7 +42,7 @@ func newConfig(opts []Option) *config {
 }
 
 // WithPropagators sets the propagators to use for Extraction and Injection
-func WithPropagators(props propagation.Propagators) Option {
+func WithPropagators(props otel.TextMapPropagator) Option {
 	return func(c *config) {
 		c.propagators = props
 	}
@@ -52,23 +51,18 @@ func WithPropagators(props propagation.Propagators) Option {
 // Returns the Attributes, Context Entries, and SpanContext that were encoded by Inject.
 func Extract(ctx context.Context, req *http.Request, opts ...Option) ([]label.KeyValue, []label.KeyValue, trace.SpanContext) {
 	c := newConfig(opts)
-	ctx = propagation.ExtractHTTP(ctx, c.propagators, req.Header)
+	ctx = c.propagators.Extract(ctx, req.Header)
 
 	attrs := append(
 		semconv.HTTPServerAttributesFromHTTPRequest("", "", req),
 		semconv.NetAttributesFromHTTPRequest("tcp", req)...,
 	)
 
-	var correlationCtxLabels []label.KeyValue
-	baggage.MapFromContext(ctx).Foreach(func(lbl label.KeyValue) bool {
-		correlationCtxLabels = append(correlationCtxLabels, lbl)
-		return true
-	})
-
-	return attrs, correlationCtxLabels, trace.RemoteSpanContextFromContext(ctx)
+	labels := otel.Baggage(ctx)
+	return attrs, (&labels).ToSlice(), trace.RemoteSpanContextFromContext(ctx)
 }
 
 func Inject(ctx context.Context, req *http.Request, opts ...Option) {
 	c := newConfig(opts)
-	propagation.InjectHTTP(ctx, c.propagators, req.Header)
+	c.propagators.Inject(ctx, req.Header)
 }