+++ instrumentation/net/http/httptrace/otelhttptrace/httptrace_test.go
@@ -24,10 +24,10 @@ import (
 	"github.com/google/go-cmp/cmp"
 
 	"go.opentelemetry.io/contrib/instrumentation/net/http/httptrace/otelhttptrace"
-	"go.opentelemetry.io/otel/api/baggage"
-	"go.opentelemetry.io/otel/api/propagation"
+	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/api/trace/tracetest"
 	"go.opentelemetry.io/otel/label"
+	"go.opentelemetry.io/otel/propagators"
 	"go.opentelemetry.io/otel/semconv"
 )
 
@@ -37,10 +37,12 @@ func TestRoundtrip(t *testing.T) {
 	var expectedAttrs map[label.Key]string
 	expectedCorrs := map[label.Key]string{label.Key("foo"): "bar"}
 
+	props := otelhttptrace.WithPropagators(otel.NewCompositeTextMapPropagator(propagators.TraceContext{}, propagators.Baggage{}))
+
 	// Mock http server
 	ts := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			attrs, corrs, span := otelhttptrace.Extract(r.Context(), r)
+			attrs, corrs, span := otelhttptrace.Extract(r.Context(), r, props)
 
 			actualAttrs := make(map[label.Key]string)
 			for _, attr := range attrs {
@@ -96,9 +98,9 @@ func TestRoundtrip(t *testing.T) {
 	err := func(ctx context.Context) error {
 		ctx, span := tr.Start(ctx, "test")
 		defer span.End()
-		ctx = baggage.ContextWithMap(ctx, baggage.NewMap(baggage.MapUpdate{SingleKV: label.Key("foo").String("bar")}))
+		ctx = otel.ContextWithBaggageValues(ctx, label.String("foo", "bar"))
 		req, _ := http.NewRequest("GET", ts.URL, strings.NewReader("foo"))
-		otelhttptrace.Inject(ctx, req)
+		otelhttptrace.Inject(ctx, req, props)
 
 		res, err := client.Do(req)
 		if err != nil {
@@ -121,7 +123,7 @@ func TestSpecifyPropagators(t *testing.T) {
 	// Mock http server
 	ts := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			_, corrs, span := otelhttptrace.Extract(r.Context(), r, otelhttptrace.WithPropagators(propagation.New(propagation.WithExtractors(baggage.DefaultHTTPPropagator()))))
+			_, corrs, span := otelhttptrace.Extract(r.Context(), r, otelhttptrace.WithPropagators(propagators.Baggage{}))
 
 			actualCorrs := make(map[label.Key]string)
 			for _, corr := range corrs {
@@ -148,9 +150,9 @@ func TestSpecifyPropagators(t *testing.T) {
 	err := func(ctx context.Context) error {
 		ctx, span := tr.Start(ctx, "test")
 		defer span.End()
-		ctx = baggage.ContextWithMap(ctx, baggage.NewMap(baggage.MapUpdate{SingleKV: label.Key("foo").String("bar")}))
+		ctx = otel.ContextWithBaggageValues(ctx, label.String("foo", "bar"))
 		req, _ := http.NewRequest("GET", ts.URL, nil)
-		otelhttptrace.Inject(ctx, req, otelhttptrace.WithPropagators(propagation.New(propagation.WithInjectors(baggage.DefaultHTTPPropagator()))))
+		otelhttptrace.Inject(ctx, req, otelhttptrace.WithPropagators(propagators.Baggage{}))
 
 		res, err := client.Do(req)
 		if err != nil {