+++ propagators/jaeger/jaeger_integration_test.go
@@ -23,7 +23,6 @@ import (
 
 	mocktracer "go.opentelemetry.io/contrib/internal/trace"
 	"go.opentelemetry.io/contrib/propagators/jaeger"
-	"go.opentelemetry.io/otel/api/propagation"
 	"go.opentelemetry.io/otel/api/trace"
 )
 
@@ -49,7 +48,6 @@ func TestExtractJaeger(t *testing.T) {
 
 	for _, tg := range testGroup {
 		propagator := jaeger.Jaeger{}
-		props := propagation.New(propagation.WithExtractors(propagator))
 
 		for _, tc := range tg.testcases {
 			t.Run(tc.name, func(t *testing.T) {
@@ -59,7 +57,7 @@ func TestExtractJaeger(t *testing.T) {
 				}
 
 				ctx := context.Background()
-				ctx = propagation.ExtractHTTP(ctx, props, req.Header)
+				ctx = propagator.Extract(ctx, req.Header)
 				resSc := trace.RemoteSpanContextFromContext(ctx)
 				if diff := cmp.Diff(resSc, tc.expected); diff != "" {
 					t.Errorf("%s: %s: -got +want %s", tg.name, tc.name, diff)